Use Netty's backpressure mechanism
Calls to TcpTransport.write() spin on the channel group waiting for a
channel that is writable. If the server gets overloaded, we let Netty's
client watermark flip the channel to non-writable and refuse to enqueue
more writes for a while. If slowing down the writing client threads for up
to 5 seconds (configurable via writeTimeout.set()) doesn't resolve the
backpressure problem, we start delivering OverloadedExceptions as usual.

The overall goal here is to reduce fragility and simplify error handling
code in high-throughput required-delivery applications. Before,
backpressure was entirely up to the user and overloading the system was
quite easy. Now, if you only have a small number of writing threads, we
should be able to soft-limit the user to reasonable latencies.

This removes the check on queue size in TcpHandler in favor of Netty's
internal backpressure mechanism.

Tests haven't been written yet.